added save folder to arguments

+++ batch.py
@@ -6,11 +6,15 @@ from pathlib import Path
 
 
 def main(seed: int, N: int, B: float, T: float, iterations: int, save_folder: Path):
+    save_file = save_folder / f'{N}_{B}_{T}' / f"{seed}.json"
+    if save_file.exists():
+        return
+    save_file.parent.mkdir(parents=True, exist_ok=True)
+
     np.random.seed(seed)
     model = TwoSpinSpherical(N, B, T)
     H_array, m_array = model.flip_many_times(iterations)
-    save_file = save_folder / f'{N}_{B}_{T}' / f"{seed}.json"
-    save_file.parent.mkdir(parents=True, exist_ok=True)
+
     with open(save_file, 'w') as f:
         data = dict(m_init=m_array[0], m_final=m_array[-1],
                     H_init=H_array[0], H_final=H_array[-1])
